% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-interface.R
\name{gbtm_engines}
\alias{gbtm_engines}
\title{Registered estimation engines}
\usage{
gbtm_engines()
}
\value{
Character vector of engine names usable as the \code{engine} argument to
\code{\link[=gbtm_fit]{gbtm_fit()}}.
}
\description{
Registered estimation engines
}
